src/tools_src/count_objects/pyproject.toml
[tool.poetry]
name = "picturas-count-objects-tool-ms"
version = "0.1.0"
description = "Picturas tool para contar objectos"
authors = ["Ana Joao Alves"]


[tool.poetry.dependencies]
python = "^3.12"
pika = "^1.3.2"
pydantic = "^2.10.3"
opencv-python = "^4.10.0"
ultralytics = "^8.3.61"
boto3 = "^1.36.2"
pillow = "^11.1.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
